refactor helper function location
