+++ helpers/__init__.py
@@ -1,1 +1,2 @@
 from .nodes import generate_nodes
+from .import_shapefiles import import_shapefiles
